requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pymysql==1.1.0
gunicorn==21.2.0
mysql-connector-python

